Support for byval aggregate arguments
The ABI for passing byval aggregates as arguments isn't
really documented, but this is what gcc seems to do:

If the aggregate fits in the available argument registers
it will be passed in those.
If it doesn't fit, it will be passed (completely) on the
stack and any arguments after the byval argument are
also passed on the stack.